Rework adapter's shutdown method to be idempotent
A test -- which I have yet to commit -- found a case where, instead of
dropping the connection, we log it as a 500. This I'll fix soon.

But something else caught my attention: different streams behave
differently -- all within specification. Combine this with accidentally
calling the shutdown method twice, and a `ClosedResourceError` would get
raised.

This does not happen with `SocketStream`s, but a bit of robustness will
not hurt.