Test main handler
There's lots of repetition -- I might refactor that later and use a
proper HTTP client -- but it does the job.

TODO: test the 504 error response. This only happens is if the TCP
handshake is too slow. I'm not sure how tricky this is to implement,
so I'll leave it for another day.